chore: useTemplateData hooks 폴더 분리
